Improve handling of context menus with GTK
When Geckofx gets displayed by GTK we can't use an SWF context menu because that
tries to track the menu so that it knows when to close it. However, because
geckofx/GTK intercepts the messages the context menu never closes.

This change translates the SWF menu items into GTK menu items and displays a
GTK menu which then get handled properly by GTK code.